% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_input.R
\name{auto_psi}
\alias{auto_psi}
\title{Set psi of the Minnesota prior}
\usage{
auto_psi(x, lags)
}
\arguments{
\item{x}{Numeric matrix with the data.}

\item{lags}{Numeric scalar. Number of lags in the model.}
}
\value{
Returns a list with the modes, minimum, and maximum values for psi.
}
\description{
Set the prior values of psi by fitting an \eqn{AR(p)} model and using the
squareroot of the innovations variance.
}
