% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_fcast_setup.R
\name{bv_cdfcast}
\alias{bv_cdfcast}
\title{Conditional forecast settings}
\usage{
bv_cdfcast(path, cond_var = NULL)
}
\arguments{
\item{path}{Numeric vector or matrix. Contains the path(s) of variable(s)
on which forecasts are conditioned on. Unrestricted future realisations
should be filled with \code{NA}. Note that not all variables can be
restricted at the same time.}

\item{cond_var}{Optional vector. Containing variable names or positions in
case \emph{path} only restricts a subset of the variables.}
}
\value{
Returns a named list of class \code{bv_cdfcast} with options for
\code{\link{bv_fcast}} or \code{\link{predict.bvar}}.
}
\description{
Provide forecast settings to \code{\link{bv_fcast}} or
\code{\link{predict.bvar}}. Allows adjustment of settings for conditional
forecasts i.e. the conditions imposed on certain variables and their
respective name or position in the dataset.
}
\examples{
# Set the path of the second variable for the first six forecast periods.
bv_cdfcast(path = c(1, 1, 1, 1, 1, 1), cond_var = 2)

# Set the path of the first and third variable at different times.
paths <- matrix(NA, nrow = 10, ncol = 2)
paths[1:5, 1] <- 1
paths[6:10, 2] <- 2
bv_cdfcast(path = paths, cond_var = c(1, 3))

# Set path for all variables (assuming 3-variable VAR).
paths <- matrix(NA, nrow = 10, ncol = 3)
paths[1:5, 1] <- 1
paths[6:10, 2] <- 2
paths[3:8, 3]  <- 3
bv_cdfcast(path = paths)
}
\seealso{
\code{\link{predict.bvar}}; \code{\link{plot.bvar_fcast}};
}
\keyword{BVAR}
\keyword{VAR}
\keyword{conditonal}
\keyword{forecasts}
\keyword{settings}
