% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_metropolis.R
\name{bv_metropolis}
\alias{bv_metropolis}
\title{Metropolis-Hastings settings}
\usage{
bv_metropolis(scale_hess = 0.01, adjust_acc = FALSE,
  acc_lower = 0.25, acc_upper = 0.35, acc_change = 0.01)
}
\arguments{
\item{scale_hess}{Numeric scalar. Scaling parameter, determining the range
of hyperparameter draws. \emph{Should be calibrated so a reasonable rate of
acceptance is reached}. Set to 0.01 by default.}

\item{adjust_acc}{Logical scalar. Whether or not to further scale the
variability of parameter draws during the burn-in phase. Turned off by
default. See details.}

\item{acc_lower}{Numeric scalar. Lower bound of the target acceptance rate.
Required if adjust_acc is set to \code{TRUE}. Defaults to 25 percent.}

\item{acc_upper}{Numeric scalar. Upper bound of the target acceptance rate.
Required if adjust_acc is set to \code{TRUE}. Defaults to 35 percent.}

\item{acc_change}{Numeric scalar. Percent change applied to the Hessian
matrix. Required if adjust_acc is set to \code{TRUE}. Defaults to 0.01.}
}
\value{
Returns a named list of class bv_metropolis with options for
\code{\link{bvar}}.
}
\description{
Function to provide settings for the Metropolis-Hastings step to
\code{\link{bvar}}. Provides options for scaling the inverse Hessian, that
is used to draw parameter proposals. Furthermore, automatic scaling to
achieve a certain acceptance rate can be performed during burn in phase.
}
\details{
If turned on, acceptance adjustment occurs only during the burn-in phase
and on every 100th iteration. Scaling is applied based on the acceptance
rate achieved for these 100 iterations. If this rate is below acc_lower the
Hessian is multiplied with \eqn{1 - acc_change}; if it is above acc_upper it
is multiplied with \eqn{1 + acc_change}.
}
\examples{
# Only adjust the scale parameter
bv_metropolis(scale_hess = 10)

# Turn on automatic scaling of the acceptance rate to [20, 40]
bv_metropolis(adjust_acc = TRUE, acc_lower = 0.2, acc_upper = 0.4)
}
