% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_ml.R
\name{bv_ml}
\alias{bv_ml}
\title{Log-posterior of a BVAR}
\usage{
bv_ml(hyper, hyper_min = -Inf, hyper_max = Inf, pars, priors, Y, X, K,
  M, N, lags, opt = FALSE)
}
\arguments{
\item{hyper}{Named numerical vector with hyperparameters for hierarchical
estimation.}

\item{hyper_min}{Optional numerical vector. Minimum values allowed for
hyperparameters. If these are breached a value of -1e18 is returned.}

\item{hyper_max}{Optional numerical vector. Maximum values allowed for
hyperparameters. If these are breached a value of -1e18 is returned.}

\item{pars}{Named numerical vector with prior parameters. Values also found
in hyper are overwritten with their hierarchical counterpart.}

\item{priors}{List created via \code{\link{bv_priors}}. Contains information
on the Minnesota prior and optional dummy priors (that are named in
\code{priors$dummy}).}

\item{Y}{Numeric \eqn{N * M} matrix.}

\item{X}{Numeric \eqn{N * K} matrix.}

\item{K}{Integer scalar. Columns of X, i.e. \eqn{M * lags + 1}.}

\item{M}{Integer scalar. Columns of Y.}

\item{N}{Integer scalar. Rows of Y, alternatively X.}

\item{lags}{Integer scalar. Number of lags in the model.}

\item{opt}{Optional logical scalar. Determines whether the return values is
a numeric scalar or a list. Used to call \code{\link{bv_ml}} in
\code{\link[stats]{optim}}.}
}
\value{
Returns a list with the following objects by default:
\itemize{
  \item log_ml - A numerical scalar with the log-posterior
  \item X, N - The respective parameters adjusted for eventual dummy
  priors. Necessary for drawing from posterior distributions.
  \item psi, sse, beta_hat, omega_inv - Further values necessary for drawing
  from posterior distributions.
}
If opt is \code{TRUE} only the value of log_ml is returned.
}
\description{
Compute the log-posterior (or log-marginal-likelihood) of a Bayesian VAR
with a Minnesota prior and optional dummy priors. Prior parameters may be
treated hierarchically. Create objects necessary for drawing from the
posterior distributions of coefficients and covariance matrix of the
residuals.
}
