% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_priors.R
\name{bv_priors}
\alias{bv_priors}
\title{Prior settings}
\usage{
bv_priors(hyper = "auto", mn = bv_mn(bv_lambda(0.2, 0.4, 0.0001, 5),
  bv_alpha(2, 0.1, 0.5), var = 10000000), ...)
}
\arguments{
\item{hyper}{Character vector. Used to specify the parameters to be treated
as hyperparameters. Can also be set to "auto" or "full" for
an automatic / full subset. Other allowed values are the Minnesota prior's
parameters "lambda", "alpha" and "psi" as well as the names of possible
dummy priors.}

\item{mn}{List of class "bv_minnesota". Options for the Minnesota prior,
set via \code{\link{bv_mn}}.}

\item{...}{Optional lists of class "bv_dummy" with options for possible
dummy priors. Must be assigned a name in the function call. Created with
\code{\link{bv_dummy}}}
}
\value{
Returns a named list of class bv_priors with options for
\code{\link{bvar}}.
}
\description{
Helper function to set up parameters and informations on priors for
\code{\link{bvar}}. Used for adjusting the parameters treated as
hyperparameters, the Minnesota prior and adding various dummy priors.
}
\examples{
# Extending hyperparameters
bv_priors(c("lambda", "alpha", "psi"))

# Adding already constructed dummy priors
# bv_priors(hyper = "auto", "soc" = soc_prior, "sur" = sur_prior)
}
\seealso{
\code{\link{bv_mn}}; \code{\link{bv_dummy}}
}
