% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_bvar.R
\name{bvar}
\alias{bvar}
\title{Hierarchical Bayesian Vector Autoregression}
\usage{
bvar(data, lags, n_draw = 10000, n_burn = 5000, n_thin = 1,
  priors = bv_priors(), metropolis = bv_metropolis(),
  fcast = bv_fcast(), irf = bv_irf(), verbose = FALSE, ...)
}
\arguments{
\item{data}{Numeric matrix or dataframe.}

\item{lags}{Integer scalar.}

\item{n_draw}{Integer scalar.}

\item{n_burn}{Integer scalar.}

\item{n_thin}{Integer scalar.}

\item{priors}{\code{bv_priors} object. See \code{\link{bv_priors}}.}

\item{metropolis}{\code{bv_metropolis} object.
See \code{\link{bv_metropolis}}.}

\item{fcast}{\code{bv_fcast} object. See \code{\link{bv_fcast}}.}

\item{irf}{\code{bv_irf} object. See \code{\link{bv_irf}}.}

\item{verbose}{Logical scalar. Whether to print intermediate results and
progress.}

\item{...}{Not used.}
}
\value{
Returns a \code{bvar} object.
}
\description{
Hierarchical Bayesian Vector Autoregression
}
