% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{check_iv}
\alias{check_iv}
\title{Check IV}
\usage{
check_iv(residuals, instrument)
}
\arguments{
\item{residuals}{Numeric matrix. During estimation, draw specific residuals
are used and rownames are inherited from data.}

\item{instrument}{Numeric vector.}
}
\value{
Returns a list of class \code{check_iv} with information on
values and indices corresponding to the intersection. Additionally the list
contains inputs provided to the function.
}
\description{
Intersects the \emph{residuals} and the \emph{instrument} based on an index
(e.g. dates) provided in rownames and names, respectively. Returns an error
 if the intersection is empty.
}
