% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/92_coef.R
\name{coef.bvar}
\alias{coef.bvar}
\alias{vcov.bvar}
\title{Coefficient and VCOV methods for Bayesian VARs}
\usage{
\method{coef}{bvar}(object, conf_bands = 0.5, companion = FALSE, ...)

\method{vcov}{bvar}(object, conf_bands = 0.5, ...)
}
\arguments{
\item{object}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{conf_bands}{Numeric vector of desired confidence bands to apply.
E.g. for bands at 5\%, 10\%, 90\% and 95\% set this to \code{c(0.05, 0.1)}.
Note that the median, i.e. \code{0.5} is always included.}

\item{companion}{Logical scalar. Whether to retrieve the companion matrix of
coefficients. See \code{\link{companion.bvar}}.}

\item{...}{Not used.}
}
\value{
Returns a numeric array of class \code{bvar_coefs} or
\code{bvar_vcovs} with values at the specified confidence bands.
}
\description{
Retrieves coefficient / variance-covariance values for Bayesian VARs
generated with \code{\link{bvar}}. Note that coefficients are available for
every stored draw and credible intervals may be set via the
\emph{conf_bands} argument.
}
\examples{
\donttest{
data <- matrix(rnorm(200), ncol = 2)
x <- bvar(data, lags = 2)

# Get coefficent values at the 10\%, 50\% and 90\% quantiles
coef(x, conf_bands = 0.10)

# Only get the median of the variance-covariance matrix
vcov(x, conf_bands = 0.5)
}
}
\seealso{
\code{\link{bvar}}; \code{\link{companion.bvar}}
}
