% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_fcast_compute.R
\name{compute_fcast}
\alias{compute_fcast}
\title{Forecast computation}
\usage{
compute_fcast(Y, K, M, N, lags, horizon, beta_comp, beta_const, sigma,
  conditional = NULL)
}
\arguments{
\item{Y}{Numeric matrix (\eqn{N * M}).}

\item{K}{Integer scalar. Columns of X, i.e. \eqn{M * lags + 1}.}

\item{M}{Integer scalar. Columns of Y.}

\item{N}{Integer scalar. Rows of Y, alternatively X.}

\item{lags}{Integer scalar. Number of lags in the model.}

\item{horizon}{Integer scalar. Specifies the horizon for which forecasts
should be computed.}

\item{beta_comp}{Numeric matrix. Posterior draw of the VAR coefficients of
the model in state space representation.}

\item{beta_const}{Numeric vector. Posterior draw of the VAR coefficients
corresponding to the constant of the model.}

\item{sigma}{Numeric matrix. Posterior draw of the vcoc-matrix of the
model.}

\item{conditional}{Not yet implemented.}
}
\value{
Returns a matrix containing forecasts for all variables in the model.
}
\description{
Compute unconditional forecasts from the VAR's posterior draws obtained via
\code{\link{draw_post}}.
}
