% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61_irf_compute.R
\name{compute_irf}
\alias{compute_irf}
\title{Impulse response draws}
\usage{
compute_irf(beta_comp, sigma, sigma_chol, M, lags, horizon, identification,
  sign_restr, fevd)
}
\arguments{
\item{beta_comp}{Numeric matrix. Posterior draw of the VAR coefficients of
the model in state space representation.}

\item{sigma}{Numeric matrix. Posterior draw of the vcoc-matrix of the
model.}

\item{sigma_chol}{Numeric matrix. Lower part of the Choleski decomposition
of sigma_draw. Calculated as \code{t(chol(sigma_draw))}.}

\item{M}{Integer scalar. Columns of \emph{X}.}

\item{lags}{Integer scalar. Number of lags in the model.}

\item{horizon}{Integer scalar. Specifies the horizon for which impulse
responses should be computed.}

\item{identification}{Logical scalar. Specifies whether or not the shocks
used for calculating the impulse should be identified. Default is
\code{TRUE}, identification will then be achieved recursively i.e. through a
Cholesky decomposition of the vcov-matrix if \emph{sign_restr} is
\code{NULL}. If set to \code{FALSE}, shocks will be unidentified.}

\item{sign_restr}{Numeric matrix. Elements inform about expected impacts
of certain shocks. Can be either \eqn{1}, \eqn{-1} or \eqn{0}, depending
on whether a positive, a negative or no contemporaneous effect of a certain
shock is expected.}

\item{fevd}{Logical scalar. Specifies whether or not forecast error variance
decompositions should be calculated.}
}
\value{
Returns impulse reponses using the posterior draws of the VAR
coefficients and the vcov-matrix of the model either to identified shocks,
using a Cholesky decomposition or sign restrictions, or unidentified ones.
}
\description{
Computes impulse responses using the posterior draws of the VAR coefficients
and vcov-matrix obtained from \code{\link{draw_post}}.
}
