% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/93_density.R
\name{density.bvar}
\alias{density.bvar}
\alias{print.bvar_density}
\alias{plot.bvar_density}
\title{Density methods for Bayesian VARs}
\usage{
\method{density}{bvar}(x, vars = NULL, ...)

\method{print}{bvar_density}(x, ...)

\method{plot}{bvar_density}(x, mar = c(2, 2, 2, 0.5), ...)
}
\arguments{
\item{x}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{vars}{Optional character vector used to subset the plot. The elements
need to match the names of hyperparameters (including \code{"ml"}). Defaults
to \code{NULL}, i.e. all variables.}

\item{...}{Not used.}

\item{mar}{Numeric vector. Margins for \code{\link[graphics]{par}}.}
}
\value{
Returns a list with outputs of \code{\link[stats]{density}}.
}
\description{
Calculates densities of hyperparameters of Bayesian VARs generated via
\code{\link{bvar}}.
}
\examples{
\donttest{
data <- matrix(rnorm(200), ncol = 2)
x <- bvar(data, lags = 2)

# Get densities of standard hyperparameters
density(x)

# Plot them
plot(density(x))

# Only get the density of the marginal likelihood
density(x, vars = "ml")
}
}
\seealso{
\code{\link{bvar}}
}
