% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_draw.R
\name{draw_post}
\alias{draw_post}
\title{BVAR posterior draws}
\usage{
draw_post(X, N = nrow(X), M = ncol(X), lags, b, psi, sse, beta_hat,
  omega_inv)
}
\arguments{
\item{X}{Numeric matrix. Possibly extended with dummy priors.}

\item{N}{Integer scalar. Rows of X.}

\item{M}{Integer scalar. Columns of X.}

\item{lags}{Integer scalar. Number of lags in the model.}

\item{b}{Numeric marix. Minnesota prior mean.}

\item{psi}{Numeric matrix. Scale of the IW prior on the residual covariance.}

\item{sse}{Numeric matrix. Squared VAR residuals.}

\item{beta_hat}{Numeric matrix.}

\item{omega_inv}{Numeric matrix.}
}
\value{
Returns a list with the following elements:
\itemize{
  \item beta_draw, sigma_draw - Draws from the posterior.
  \item sigma_chol - The Cholesky factorisation of sigma_draw. Calculated
  as \code{t(chol(sigma_draw))}.
}
}
\description{
Draw beta and sigma from the posterior of a BVAR.
}
