% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_dummy.R
\name{dummy}
\alias{dummy}
\title{Prior settings}
\usage{
dummy(mode = 1, min = 0.0001, max = 5, ...)
}
\arguments{
\item{mode}{Numeric scalar. Mode or the like of the parameter.}

\item{min}{Numeric scalar. Minimum allowed value.}

\item{max}{Numeric scalar. Maximum allowed value.}

\item{...}{Other possible parameters such as sd or fun.}
}
\value{
Returns a list of class bv_dummy.
}
\description{
Prior settings
}
