% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63_fevd.R
\name{fevd}
\alias{fevd}
\title{Forecast error variance decompostions draws}
\usage{
fevd(irf_comp, M, horizon)
}
\arguments{
\item{irf_comp}{Numeric matrix. Contains a draw of impulse responses
obtained from \code{\link{compute_irf}}.}

\item{M}{Integer scalar. Columns of X.}

\item{horizon}{Integer scalar. Horizon of impulse responses and FEVDs.}
}
\value{
Returns a numeric matrix of FEVDs.
}
\description{
Computes Forecast error variance decompostions (FEVDs) using the impulse
response draws obtained via \code{\link{compute_irf}}.
}
