% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/95_fitted.R
\name{fitted.bvar}
\alias{fitted.bvar}
\alias{residuals.bvar}
\alias{plot.bvar_resid}
\title{Fitted and residual methods for Bayesian VARs}
\usage{
\method{fitted}{bvar}(object, conf_bands = 0.5, ...)

\method{residuals}{bvar}(object, conf_bands = 0.5, ...)

\method{plot}{bvar_resid}(x, vars = NULL, mar = c(2, 2, 2, 0.5), ...)
}
\arguments{
\item{object}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{conf_bands}{Numeric vector of desired confidence bands to apply.
E.g. for bands at 5\%, 10\%, 90\% and 95\% set this to \code{c(0.05, 0.1)}.
Note that the median, i.e. \code{0.5} is always included.}

\item{...}{Graphical parameters for \code{\link[graphics]{par}}.}

\item{x}{Object of class \code{bvar_fitted} / \code{bvar_resid}.}

\item{vars}{Optional numeric vector. Used to subset the plot to certain
variables by position. Defaults to \code{NULL}, i.e. all variables.}

\item{mar}{Numeric vector. Margins for \code{\link[graphics]{par}}.}
}
\value{
Returns a numeric array of class \code{bvar_fitted} or
\code{bvar_resid} with values at the specified confidence bands.
}
\description{
Calculates fitted or residual values for Bayesian VARs generated with
\code{\link{bvar}}.
}
\examples{
\donttest{
data <- matrix(rnorm(200), ncol = 2)
x <- bvar(data, lags = 2)

# Get fitted values and adjust confidence bands to 10\%, 50\% and 90\%
fitted(x, conf_bands = 0.10)

# Get residuals and plot them
plot(residuals(x))
}
}
\seealso{
\code{\link{bvar}}
}
