% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fred_transform}
\alias{fred_transform}
\title{FRED transformation and subsetting helper}
\usage{
fred_transform(
  data,
  type = c("qd", "md"),
  transform = TRUE,
  na.rm = TRUE,
  lag = 1,
  scale = 100
)
}
\arguments{
\item{data}{A \code{data.frame} with FRED-QD or FRED-MD time series. The
column names are used to find the correct transformation.}

\item{type}{Character scalar. Specify whether \emph{data} stems from the
FRED-QD or FRED-MD database.}

\item{transform}{Logical scalar. Whether to apply transformations.}

\item{na.rm}{Logical scalar. Whether to subset to rows without any
\code{NA} values. A warning is thrown if rows are non-sequential.}

\item{lag}{Integer scalar. Number of lags to apply in the transformation
when taking differences. Defaults to 1.}

\item{scale}{Numeric scalar. Factor by which to scale up values after taking
differences. Defaults to 100.}
}
\value{
Returns a \code{data.frame} object.
}
\description{
Apply transformations given by FRED-MD or FRED-QD and generate rectangular
subsets. See \code{\link{fred_qd}} for information on data and the details
section for information on the transformations.
}
\details{
FRED-QD and FRED-MD include a transformation code for every variable. All
codes are provided in \code{system.file("fred_trans.rds", package = "BVAR")}.
The transformation codes are as follows:
\enumerate{
  \item \code{1} - no transformation;
  \item \code{2} - first differences - \eqn{\delta x_t};
  \item \code{3} - second differences - \eqn{\delta^2 x_t};
  \item \code{4} - log transformation - \eqn{log(x_t)};
  \item \code{5} - log differences - \eqn{\delta^2 log(x_t)};
  \item \code{6} - percent change differences -
  \eqn{\delta x_t / x_{t-1} - 1};
}
Note that the transformation codes of FRED-MD and FRED-QD may differ for
the same series.
}
\examples{
# Transform a subset of FRED-QD
fred_transform(fred_qd[, c("GDPC1", "INDPRO", "FEDFUNDS")])

# Transform all of FRED-MD
\dontrun{
fred_transform(fred_md, type = "md")
}
}
\seealso{
\code{\link{fred_qd}};
}
