% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{gamma_coef}
\alias{gamma_coef}
\title{Compute gamma coefficients}
\usage{
gamma_coef(mode, sd)
}
\arguments{
\item{mode}{Numeric scalar.}

\item{sd}{Numeric scalar.}
}
\value{
Returns a list with shape \emph{k} and scale paramter \emph{theta}.
}
\description{
Compute the shape \emph{k} and scale \emph{theta} of a gamma
distribution via mode and standard deviation.
}
