% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{get_var_set}
\alias{get_var_set}
\title{Get a subset of variables}
\usage{
get_var_set(vars, variables, M)
}
\arguments{
\item{vars}{Vector of variables to subset to. Numeric or character.}

\item{variables}{Character vector of all variable names. Required if vars
is provided as character vector.}

\item{M}{Integer scalar. Count of all variables.}
}
\value{
Returns a numeric vector with the positions of desired variables.
}
\description{
Helper functions to aid with variable selection in \code{\link{bv_plot_irf}}
and \code{\link{bv_plot_fcast}}.
}
\examples{
# Assuming the variables are named.
bvar:::get_var_set("fx_rate", variables = c("gdp_pc", "fx_rate"))

# Find via position
bvar:::get_var_set(c(1, 3), M = 3)

# Get the full set
bvar:::get_var_set(NULL, M = 3)
}
