% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/64_irf_method.R, R/65_irf_print.R
\name{irf.bvar}
\alias{irf.bvar}
\alias{irf.bvar_irf}
\alias{fevd.bvar}
\alias{fevd.bvar_irf}
\alias{irf}
\alias{fevd}
\alias{print.bv_irf}
\alias{print.bvar_irf}
\alias{print.bvar_fevd}
\alias{summary.bvar_irf}
\title{Impulse response and forecast error methods for Bayesian VARs}
\usage{
\method{irf}{bvar}(x, ..., conf_bands, n_thin = 1L)

\method{irf}{bvar_irf}(x, conf_bands, ...)

\method{fevd}{bvar}(x, ..., conf_bands = 0.5, n_thin = 1L)

\method{fevd}{bvar_irf}(x, conf_bands = 0.5, ...)

irf(x, ...)

fevd(x, ...)

\method{print}{bv_irf}(x, ...)

\method{print}{bvar_irf}(x, ...)

\method{print}{bvar_fevd}(x, digits = 4L, ...)

\method{summary}{bvar_irf}(object, vars_impulse = NULL,
  vars_response = NULL, digits = 2L, ...)
}
\arguments{
\item{x}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{...}{A \code{bv_fcast} object or parameters to be fed into
\code{\link{bv_fcast}}. Contains settings for the forecast.}

\item{conf_bands}{Numeric vector of desired confidence bands to apply.
E.g. for bands at 5\%, 10\%, 90\% and 95\% set this to \code{c(0.05, 0.1)}.}

\item{n_thin}{Integer scalar. Every \emph{n_thin}'th draw in \emph{x} is used
for forecasting, others are dropped. Defaults to the maximum number - i.e.
the number of saved draws in \emph{x}.}

\item{digits}{Integer scalar. Fed to \code{\link[base]{round}} and applied to
numeric outputs (i.e. the quantiles).}

\item{object}{Object of class \code{bvar_irf}.}

\item{vars_impulse}{Optional numeric or character vector. Used to subset the
plot's impulses to certain variables by position or name (must be available).
Defaults to \code{NULL}, i.e. all variables.}

\item{vars_response}{Optional numeric or character vector. Used to subset the
plot's responses to certain variables by position or name (must be
available). Defaults to \code{NULL}, i.e. all variables.}
}
\value{
Returns a list of class \code{bvar_irf} including IRFs, optionally
FEVDs, and desired confidence bands. See \code{\link{bvar}}.
Note that \code{fevd} only returns a numeric array of FEVDs and desired
confidence bands.

Returns an array with the desired impulse response quantiles
invisibly.
}
\description{
Retrieves / calculates impulse response functions (IRFs) / forecast error
variance decompositions (FEVDs) for Bayesian VARs generated via
\code{\link{bvar}}. If the object is already present and no settings are
supplied it is simply retrieved. Otherwise it will be calculated ex-post.
Note that FEVDs require the presence / calculation of IRFs. May also be used
to update confidence bands.

Print and return quantiles of impulse response functions from
\code{\link{bvar}}.
}
\examples{
\donttest{
data <- matrix(rnorm(200), ncol = 2)
x <- bvar(data, lags = 2)

# Access forecast and update new confidence bands
predict(x, conf_bands = 0.01)

# Compute and store a longer forecast
x$fcast <- predict(x, horizon = 24L)

# Speed up by lowering draws and use bv_fcast() to set options
predict(x, bv_fcast(24L), n_thin = 10L)

# Update the confidence bands
x$fcast <- predict(x$fcast, conf_bands = c(0.05, 0.16))

# Use new data to calculate a prediction
predict(x, newdata = matrix(rnorm(200), ncol = 2))
}
}
