% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{lag_var}
\alias{lag_var}
\title{Lag a matrix of time series}
\usage{
lag_var(x, lags)
}
\arguments{
\item{x}{Matrix (\eqn{N * M}) to lag.}

\item{lags}{Numeric scalar. Number of lags to create.}
}
\value{
Returns an \eqn{N * (M * lags)} matrix with consecutive lags at the
right.
}
\description{
Compute a lagged version of matrix to be used on vector autoregressions.
Multiple lags are added side by side.
}
