% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mpi.R
\docType{data}
\name{mpi}
\alias{mpi}
\title{Monetary Policy Instrument}
\format{
A \code{numeric vector} object with dates as rownames.
}
\source{
\url{https://github.com/riccardo-degasperi/info-policy-surprises}
}
\description{
We make use of time series at monthly frequency of the high-frequency
instruments for monetary policy and information shocks as defined in
Miranda-Agrippino and Ricco (2021) and extended to 2018:12 in
Degasperi and Ricco (2021).
The monetary policy instrument that is provided here is licensed
under a modified CC-BY 4.0 license that can be found in the provided
\emph{LICENSE} file. #' For further details see
\url{https://github.com/riccardo-degasperi/info-policy-surprises}.
We would like to thank Silvia Miranda-Agrippino, Giovanni Ricco and
Riccardo Degasperi for creating, updating and making the instrument available.
}
\references{
Miranda-Agrippino, S., & Ricco, G. (2021). The transmission of monetary
  policy shocks. \emph{American Economic Journal: Macroeconomics},
   \bold{13(3)}, 74-107, doi:10.1257/mac.20180124 \cr
   Degasperi, R. and Ricco, G., 2021.
   Information and policy shocks in monetary surprises. Working Paper.
}
\keyword{instrument}
\keyword{monetary}
\keyword{policy;}
\keyword{shocks;}
