% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{name_pars}
\alias{name_pars}
\title{Create parameter names}
\usage{
name_pars(x, M)
}
\arguments{
\item{x}{Character vector with names of all relevant paramters.}

\item{M}{Integer scalar with the number of columns in the data.}
}
\value{
Returns a character vector of parameter names.
}
\description{
Function to help name prior parameters. Accounts for multiple occurences
of psi when \eqn{M > 1} by adding sequential numbers.
}
\examples{
bvar:::name_pars(c("lambda", "alpha"))
bvar:::name_pars(c("lambda", "psi"), M = 3)
}
