% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_input.R
\name{num_check}
\alias{num_check}
\alias{int_check}
\title{Check a numeric scalar}
\usage{
num_check(x, min = 0, max = Inf,
  msg = "Please check the integer parameters.", fun = as.numeric)

int_check(x, min = 0, max = Inf,
  msg = "Please check the integer parameters.")
}
\arguments{
\item{x}{Numeric scalar to check.}

\item{min}{Numeric scalar. Minimum value of x.}

\item{max}{Numeric scalar. Maximum value of x.}

\item{msg}{String fed to \code{\link[base]{stop}} if an error occurs.}

\item{fun}{Function to apply to x before returning.}
}
\value{
Returns fun(x).
}
\description{
Function to check whether an object is properly bounded and coercible to a
a numeric value.
}
