% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/85_plot_fcast.R
\name{plot.bvar_fcast}
\alias{plot.bvar_fcast}
\alias{bv_plot_fcast}
\title{Forecast plot}
\usage{
\method{plot}{bvar_fcast}(x, conf_bands = 0.16, variables = NULL,
  vars = NULL, mar = c(2, 2, 2, 0.5), orientation = c("vertical",
  "horizontal"), ...)

bv_plot_fcast(x, conf_bands = 0.16, variables = NULL, vars = NULL,
  mar = c(2, 2, 2, 0.5), orientation = c("vertical", "horizontal"),
  ...)
}
\arguments{
\item{x}{A \code{bvar} / \code{bvar_fcast} object, obtained from
\code{\link{bvar}}.}

\item{conf_bands}{Numeric vector of desired confidence bands. For 5\%, 10\%,
90\% and 95\% bands set this to \code{c(0.05, 0.1)}.}

\item{variables}{Optional character vector. Names of all variables in the
object. Taken from \code{x$variables} if available.}

\item{vars}{Optional numeric or character vector. Used to subset the plot to
certain variables by position or name (\code{variables} must be available).
Defaults to \code{NULL}, i.e. all variables.}

\item{mar}{Numeric vector with margins for \code{\link[graphics]{par}}.}

\item{orientation}{String indicating the orientation of the plots. Defaults
to "v" (i.e. "vertical"); may be set to "h" (i.e. "horizontal").}

\item{...}{Other graphical parameters for \code{\link[graphics]{par}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Plot method for forecasts obtained from \code{\link{bvar}}. Plot forecasts
for all or a subset of the available variables.
}
\examples{
\dontrun{
# x <- bvar()

# Plot forecasts for all available variables
bv_plot_fcast(x)

# Subset to variables in positions 1, 2 and 4 via position and name
bv_plot_fcast(x, vars = c(1, 2, 4))
bv_plot_fcast(x,
  variables = c("gdp", "flux", "cpi", "capacitor"),
  vars = c("gdp", "flux", "capacitor")
)

# Use the method to plot and adjust confidence bands and orientation
plot(x$fcast, conf_bands = c(0.01, 0.05), orientation = "h")
}
}
