% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/85_plot_fcast.R
\name{plot_fcast}
\alias{plot_fcast}
\title{Forecast plot}
\usage{
plot_fcast(x, variables, pos, col, mar, mfrow, ...)
}
\arguments{
\item{x}{Numeric array (3-dimensional) with data to plot. The first
dimension contains quantiles, the seconds paths and the third variables.}

\item{variables}{Character vector with the names of variables.}

\item{pos}{Integer vector. Positions of the variables to plot.}

\item{col}{Character vector. Colours to feed to \code{\link[stats]{ts.plot}}.}

\item{mar}{Numeric vector with margins for \code{\link[graphics]{par}}.}

\item{mfrow}{Numeric vector with layout for \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters for \code{\link[graphics]{par}}.}
}
\description{
Forecast plot
}
