% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/84_plot_hyper.R
\name{plot_hyper}
\alias{plot_hyper}
\title{Hyperparameter plot helper}
\usage{
plot_hyper(x, name, fun = c(plot_trace, plot_dens), ...)
}
\arguments{
\item{x}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{name}{String with the name of the parameter to plot.}

\item{fun}{Function to use for plotting.}

\item{...}{Further \code{bvar} objects to include in the plot.}
}
\value{
Returns x invisibly.
}
\description{
Helper function to provide input checks, prepare data etc. for
hyperparameter plots.
}
