% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/86_plot_irf.R
\name{plot_irf}
\alias{plot_irf}
\title{Impulse response plot}
\usage{
plot_irf(x, variables, pos_imp, pos_res, col, mar, mfrow, ...)
}
\arguments{
\item{x}{Numeric array (4-dimensional) with data to plot. The first
dimension contains quantiles, the second responses, the third paths and the
fourth impulses.}

\item{variables}{Character vector with the names of variables.}

\item{pos_imp}{Integer vector. Positions of the impulse variables to plot.}

\item{pos_res}{Integer vector. Positions of the response variables to plot.}

\item{col}{Character vector. Colours to feed to \code{\link[stats]{ts.plot}}.}

\item{mar}{Numeric vector with margins for \code{\link[graphics]{par}}.}

\item{mfrow}{Numeric vector with layout for \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters for \code{\link[graphics]{par}}.}
}
\description{
Impulse response plot
}
