% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/84_plot_hyper.R
\name{plot_trace}
\alias{plot_trace}
\alias{plot_dens}
\title{Trace & density plot}
\usage{
plot_trace(x, name = NULL, bounds = NULL, dots = list())

plot_dens(x, name = NULL, bounds = NULL, dots = list())
}
\arguments{
\item{x}{Numeric vector to plot.}

\item{name}{Optional string with the plotted parameter's name.}

\item{bounds}{Optional numeric vector plotted horizontally via
\code{\link[graphics]{abline}}.}

\item{dots}{Optional list of numeric vectors to add to the plot.}
}
\description{
Trace & density plot
}
