% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_predict.R, R/55_fcast_print.R
\name{predict.bvar}
\alias{predict.bvar}
\alias{predict.bvar_fcast}
\alias{print.bvar_fcast}
\alias{summary.bvar_fcast}
\title{Predict method for Bayesian VARs}
\usage{
\method{predict}{bvar}(object, ..., conf_bands, n_thin = 1L, newdata)

\method{predict}{bvar_fcast}(object, conf_bands, ...)

\method{print}{bvar_fcast}(x, vars = NULL, ...)

\method{summary}{bvar_fcast}(object, vars = NULL, digits = 2L, ...)
}
\arguments{
\item{object}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{...}{A \code{bv_fcast} object or parameters to be fed into
\code{\link{bv_fcast}}. Contains settings for the forecast.}

\item{conf_bands}{Numeric vector of desired confidence bands to apply.
E.g. for bands at 5\%, 10\%, 90\% and 95\% set this to \code{c(0.05, 0.1)}.}

\item{n_thin}{Integer scalar. Every \emph{n_thin}'th draw in \emph{object} is used
for forecasting, others are dropped. Defaults to the maximum number - i.e.
the number of saved draws in \emph{object}.}

\item{newdata}{Optional numeric matrix or dataframe. Used to base the
prediction on. Fitted values are used by default.}

\item{x}{Object of class \code{bvar_fcast}.}

\item{vars}{Optional numeric or character vector. Used to subset the output
to certain variables by position or name (must be available). Defaults to
\code{NULL}, i.e. all variables.}

\item{digits}{Integer scalar. Fed to \code{\link[base]{round}} and applied to
numeric outputs (i.e. the quantiles).}
}
\value{
Returns a list of class \code{bvar_fcast} including forecasts
and desired confidence bands. See \code{\link{bvar}}.

Returns an array with the desired forecast quantiles invisibly.
}
\description{
Retrieves / calculates forecasts for Bayesian VARs generated via
\code{\link{bvar}}. If a forecast is already present and no settings are
supplied it is simply retrieved. If no forecast is present or settings are
provided one will be calculated ex-post. May also be used to update
confidence bands.

Print and return quantiles of forecasts from \code{\link{bvar}}.
}
\examples{
\donttest{
data <- matrix(rnorm(200), ncol = 2)
x <- bvar(data, lags = 2)

# Access forecast and update new confidence bands
predict(x, conf_bands = 0.01)

# Compute and store a longer forecast
x$fcast <- predict(x, horizon = 24L)

# Speed up by lowering draws and use bv_fcast() to set options
predict(x, bv_fcast(24L), n_thin = 10L)

# Update the confidence bands
x$fcast <- predict(x$fcast, conf_bands = c(0.05, 0.16))

# Use new data to calculate a prediction
predict(x, newdata = matrix(rnorm(200), ncol = 2))
}
}
