% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_method.R, R/95_method_fcast.R,
%   R/96_method_irf.R
\name{print.bvar}
\alias{print.bvar}
\alias{print.bv_fcast}
\alias{print.bvar_fcast}
\alias{print.bv_irf}
\alias{print.bvar_irf}
\title{BVAR print methods}
\usage{
\method{print}{bvar}(x, ...)

\method{print}{bv_fcast}(x, ...)

\method{print}{bvar_fcast}(x, ...)

\method{print}{bv_irf}(x, ...)

\method{print}{bvar_irf}(x, ...)
}
\arguments{
\item{x}{A \code{bvar} / \code{bvar_...} object, obtained from
\code{\link{bvar}} or constructor functions.}

\item{...}{Not used.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for \code{bvar} objects. Provide a quick overview of results.
The method for impulse response functions provides information on
identification, horizon, et cetera; The forecast method on the horizon.
Further methods for \code{bv_priors}, \code{bv_metropolis}, et cetera are
in the works.
}
