% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/95_method_fcast.R
\name{print_fcast}
\alias{print_fcast}
\title{Forecast print method}
\usage{
print_fcast(x, ...)
}
\arguments{
\item{x}{A \code{bv_fcast} object.}

\item{...}{Not used.}
}
\description{
Forecast print method
}
