% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96_method_irf.R
\name{print_irf}
\alias{print_irf}
\title{Impulse response print method}
\usage{
print_irf(x, ...)
}
\arguments{
\item{x}{A \code{bv_irf} object.}

\item{...}{Not used.}
}
\description{
Impulse response print method
}
