% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62b_proxy_var.R
\name{proxy_svar}
\alias{proxy_svar}
\title{Proxy SVAR}
\usage{
proxy_svar(residuals, instrument)
}
\arguments{
\item{residuals}{Numeric matrix. During estimation, draw specific residuals
are used.}

\item{instrument}{Numeric vector containing the instrument. Possibly
shortened beforehand to match the length of residuals.}
}
\value{
Returns a list with the following elements:
\itemize{
  \item \code{impact} - Contemporaneous transmission coefficients: Bzero
  \item \code{Gamma}  - Estimated correlation between shock and instrument
  \item \code{f_stat} - First stage F statistic
  \item \code{lambda} - Reliability of instrument
}
}
\description{
Performs identification using external \emph{instrument}. At this stage the
\emph{instrument} and \emph{residuals} need to have equal number of
observations. Transformations such as intersecting the instrument and
residuals need to be done before calling this function.
}
