% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_input.R
\name{quantile_check}
\alias{quantile_check}
\title{Generate quantiles}
\usage{
quantile_check(conf_bands)
}
\arguments{
\item{conf_bands}{Numeric vector of probabilities i.e. between 0 and 1.}
}
\value{
Returns a sorted vector of quantiles created from conf_bands.
}
\description{
Check and create a given vector of confidence bands and create suitable
quantiles from it.
}
\examples{
bvar:::quantile_check(c(0.1, 0.16))
}
