% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{set_dates}
\alias{set_dates}
\title{Set dates}
\usage{
set_dates(data, start, frequency)
}
\arguments{
\item{data}{Numeric vector, matrix or a data.frame. Note that observations
are expected to be ordered from earliest to latest, and variables in the
columns.}

\item{start}{Starting date of the data in form "YYYY-MM-DD".}

\item{frequency}{Frequency of the data: "year", "quarter", "month", "week" or "day".}
}
\value{
Returns data with rownames containing sequence of dates as character.
}
\description{
Changes rownames to a character string containing dates. Common format of
dates or other index is needed if the length of the \emph{instrument} and
reduced form residuals differ. In such case, the identification is based on
a common subset of residuals and the instrument.
}
