% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_aux.R
\name{set_gray}
\alias{set_gray}
\title{Credible interval colour vector}
\usage{
set_gray(P)
}
\arguments{
\item{P}{Integer scalar. Number of bands to plot.}
}
\value{
Returns a character vector of colours.
}
\description{
Create a character vector of colours for time series with credible
intervals, e.g. \code{\link{bv_plot_irf}} and \code{\link{bv_plot_fcast}}.
The central element is set to black, the rest to darkgray.
}
\examples{
# set_gray(3)
}
