% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62_sign_restr.R
\name{sign_restr}
\alias{sign_restr}
\title{Sign restriction algorithm}
\usage{
sign_restr(sigma_chol, sign_restr, M)
}
\arguments{
\item{sigma_chol}{Numeric matrix. Lower part of the Cholesky decomposition
of the posterior draw of the vcov-matrix of the system.}

\item{sign_restr}{Numeric matrix. Elements inform about expected impacts
of certain shocks. Can be either \eqn{1}, \eqn{-1} or \eqn{0}, depending
on whether a positive, a negative or no contemporaneous effect of a certain
shock is expected.}

\item{M}{Integer scalar. Columns of X.}
}
\value{
Returns a matrix used as shock for computations of impulse responses
identified via sign restrictions.
}
\description{
Draws orthogonal matrices and checks whether sign restrictions are fulfilled
until a suitable matrix is found. Is called by \code{\link{compute_irf}}.
Throws an error if no suitable restrictions can be found.
}
