% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vardata.R
\name{var_data_prepare}
\alias{var_data_prepare}
\title{Title}
\usage{
var_data_prepare(
  data,
  lags = 1,
  const = T,
  trend = F,
  trend_qua = F,
  ex = NULL,
  ex_lag = NULL,
  xlags = 0
)
}
\arguments{
\item{xlags}{}
}
\description{
Title
}
