% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_fcast.R
\name{bv_fcast}
\alias{bv_fcast}
\title{Forecast settings}
\usage{
bv_fcast(horizon = 12, conditional = FALSE)
}
\arguments{
\item{horizon}{Integer scalar. Horizon for which forecasts should be
computed.}

\item{conditional}{Logical scalar. Not yet implemented.}
}
\value{
Returns a named list of class \code{bv_fcast} with options for
\code{\link{bvar}}.
}
\description{
Provide forecast settings to \code{\link{bvar}}. Allows adjusting the horizon
of the forecast.
}
\details{
As of now only unconditional forecasts are implemented and conditional must
be set to FALSE. Conditional forecasts will require further settings, such
as variable paths.
}
\examples{
# Set the forecast-horizon to 20 time periods
bv_fcast(horizon = 20)
}
