% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_irf_setup.R
\name{bv_irf}
\alias{bv_irf}
\title{Impulse response settings and identification}
\usage{
bv_irf(
  horizon = 12,
  fevd = FALSE,
  identification = TRUE,
  sign_restr = NULL,
  sign_lim = 1000,
  instrument = NULL,
  proxyvar = NULL,
  manual_matching = FALSE
)
}
\arguments{
\item{horizon}{Integer scalar. The horizon for which impulse responses
(and FEVDs) should be computed. Note that the first period corresponds to
impacts i.e. contemporaneous effects.}

\item{fevd}{Logical scalar. Whether or not forecast error variance
decompositions should be calculated.}

\item{identification}{Logical scalar. Whether or not the shocks used for
calculating impulses should be identified. Defaults to \code{TRUE}, i.e.
identification via Cholesky decomposition of the VCOV-matrix unless
\emph{sign_restr} or \emph{instrument} is provided.}

\item{sign_restr}{Elements inform about expected impacts
of certain shocks. Can be either \eqn{1}, \eqn{-1} or \eqn{0} depending
on whether a positive, a negative or no contemporaneous effect of a
certain shock is expected. Elements set to \eqn{NA} indicate that there are
no particular expectations for the contemporaneous effects. The default
value is \code{NULL}. Note that in order to be fully identified at least
\eqn{M * (M - 1) / 2} restrictions have to be set and a maximum of
\eqn{M - j} zero restrictions can be imposed on the \eqn{j}'th column.}

\item{sign_lim}{Integer scalar. Maximum number of tries to find suitable
matrices to for fitting sign or zero and sign restrictions.}

\item{instrument}{Numeric vector. If provided, the identification is performed using proxy
SVAR. Since the residuals and instrument need not to have equal length in general,
user is expected to provide a common index (rownames) for the data and the
instrument. See examples and helper function \emph{set_dates}. This can be
avoided by setting \emph{manual_matching} to TRUE.}

\item{proxyvar}{character string. Variable for which instrument is provided.
Needs to correspond to one of the column names of the data used in
\emph{bvar}.}

\item{manual_matching}{If set to TRUE, user is not expected to specify common
index for the data and instrument. Instead the length of instrument needs to
match the length of residuals (= length(data inputted in bvar) - lags).}
}
\value{
Returns a named list of class \code{bv_irf} with options for
\code{\link{bvar}}, \code{\link{irf.bvar}} or \code{\link{fevd.bvar}}.
}
\description{
Provides settings for the computation of impulse responses to
\code{\link{bvar}}, \code{\link{irf.bvar}} or \code{\link{fevd.bvar}}. Allows
setting the horizon for which impulse responses should be computed, whether
or not forecast error variance decompositions (FEVDs) should be included
as well as if and what kind of identification should be used. See the Details
section for further information on identification. Identification can be
achieved via Cholesky decomposition, sign restrictions (Rubio-Ramirez,
Waggoner and Zha, 2010), combination of zero and sign restrictions (Arias,
Rubio-Ramirez and Waggoner, 2018) or using an external instrument
(Stock and Watson 2012, Mertens and Ravn 2013)
}
\details{
Identification can be performed via Cholesky decomposition, sign
restrictions, zero and sign restrictions or using external instrument.
The algorithm for generating suitable sign restrictions follows
Rubio-Ramirez, Waggoner and Zha (2010), while the one for zero and sign
restrictions follows Arias, Rubio-Ramirez and Waggoner (2018).
Note the possiblity of finding no suitable zero/sign restrictions.
The proxy SVAR identification follows the implementation in
Miranda-Agrippino and Ricco (2021).
}
\examples{
# Set impulse responses to a horizon of 20 time periods and enable FEVD
# (Identification is performed via Cholesky decomposition)
bv_irf(horizon = 20, fevd = TRUE)

# Set up structural impulse responses using sign restrictions
signs <- matrix(c(1, NA, NA, -1, 1, -1, -1, 1, 1), nrow = 3)
bv_irf(sign_restr = signs)

# Set up structural impulse responses using zero and sign restrictions
zero_signs <- matrix(c(1, 0, NA, -1, 1, 0, -1, 1, 1), nrow = 3)
bv_irf(sign_restr = zero_signs)

# Set up structural impulse responses using external instrument with length
corresponding to reduced form residuals
bv_irf(instrument = istrument)

# Prepare to estimate unidentified impulse responses
bv_irf(identification = FALSE)

# Prepare to estimate SVAR-IV using an indexed instrument for monetary policy shocks
mpi_named <- readRDS(file = "./data/instrument_MAR21.rds")
bv_irf(instrument = mpi_named, manual_matching = FALSE,
       proxyvar = "GS1")

# Prepare to estimate SVAR-IV using an instrument for monetary policy shocks
# (where its length is equal to the length of the residuals)
mpi_exact <- readRDS(file = "./data/instrument_MAR21.rds")
bv_irf(instrument = mpi_exact, manual_matching = TRUE,
       proxyvar = "GS1")

}
\references{
Rubio-Ramirez, J. F. and Waggoner, D. F. and Zha, T. (2010) Structural
  Vector Autoregressions: Theory of Identification and Algorithms for
  Inference. \emph{The Review of Economic Studies}, \bold{77}, 665-696,
  doi:10.1111/j.1467-937X.2009.00578.x.
  Arias, J.E. and Rubio-Ramirez, J. F. and Waggoner, D. F. (2018)
  Inference Based on Structural Vector Autoregressions Identifiied with
  Sign and Zero Restrictions: Theory and Applications.
  \emph{Econometrica}, \bold{86}, 2, 685-720,
  doi:10.3982/ECTA14468.
  Miranda-Agrippino, S., & Ricco, G. (2021). The transmission of monetary
  policy shocks. \emph{American Economic Journal: Macroeconomics},
   \bold{13(3)}, 74-107, doi:10.1257/mac.20180124
   Stock, J. H., and Watson, M. W. (2012). Disentangling the Channels of the
   2007-2009 Recession (No. w18094) \emph{National Bureau of Economic
   Research}, doi:10.3386/w18094.
   Mertens, K., and Ravn, M. O. (2013). The dynamic effects of personal and
   corporate income tax changes in the United States. \emph{American
   economic review}, \bold{103(4)}, 1212-1247,
   doi:10.1257/aer.103.4.1212.
}
\seealso{
\code{\link{irf.bvar}}; \code{\link{plot.bvar_irf}}
}
\keyword{BVAR}
\keyword{fevd}
\keyword{irf}
\keyword{settings}
