% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_priors.R
\name{bv_priors}
\alias{bv_priors}
\title{Prior settings}
\usage{
bv_priors(hyper = "auto", mn = bv_mn(bv_lambda(0.2, 0.4, 0.0001, 5),
  bv_alpha(2, 0.1, 0.5), var = 10000000), ...)
}
\arguments{
\item{hyper}{Character vector. Used to specify the parameters to be treated
as hyperparameters. May also be set to \code{"auto"} or \code{"full"} for
an automatic / full subset. Other allowed values are the Minnesota prior's
parameters \code{"lambda"}, \code{"alpha"} and \code{"psi"} as well as the
names of additional dummy priors included via \emph{...}.}

\item{mn}{List of class \code{"bv_minnesota"}. Options for the Minnesota
prior, set via \code{\link{bv_mn}}.}

\item{...}{Optional lists of class \code{"bv_dummy"} with options for
dummy priors. \emph{Must be assigned a name in the function call}. Created
with \code{\link{bv_dummy}}.}
}
\value{
Returns a named list of class \code{bv_priors} with options for
\code{\link{bvar}}.
}
\description{
Function to provide priors and their parameters to \code{\link{bvar}}. Used
for adjusting the parameters treated as hyperparameters, the Minnesota prior
and adding various dummy priors through the ellipsis.
}
\examples{
# Extending hyperparameters to the full Minnesota prior
bv_priors(c("lambda", "alpha", "psi"))

# Adding already constructed dummy priors from `bv_dummy()`
bv_priors(hyper = "auto", soc = soc, sur = sur)
}
\seealso{
\code{\link{bv_mn}}; \code{\link{bv_dummy}}
}
