% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVAR-package.R
\docType{package}
\name{BVAR-package}
\alias{BVAR-package}
\title{BVAR}
\description{
Toolkit for the estimation of hierarchical Bayesian vector autoregressions.
Implements hierarchical prior selection for conjugate priors in the fashion
of Giannone, Lenza & Primiceri (2015). Allows for the computation of impulse
responses and forecasts and provides functionality for assessing results.
}
\references{
Giannone, D., Lenza, M., & Primiceri, G. E. (2015). Prior Selection for Vector Autoregressions. Review of Economics and Statistics, 97, 436-451. \url{https://doi.org/10.1162/REST_a_00483}.
}
