% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_bvar.R
\name{bvar}
\alias{bvar}
\title{Hierarchical Bayesian Vector Autoregression}
\usage{
bvar(
  data,
  lags,
  n_draw = 5000L,
  n_burn = 1000L,
  n_thin = 1L,
  priors = bv_priors(),
  mh = bv_mh(),
  fcast = NULL,
  irf = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{Numeric matrix or dataframe. Note that observations need to be
ordered from earliest to latest one.}

\item{lags}{Integer scalar. Number of lags to apply to the data.}

\item{n_draw, n_burn}{Integer scalar. Define the number of iterations
to (a) cycle through and (b) burn at the start.}

\item{n_thin}{Integer scalar. Provides the option of reducing the number of
stored iterations to every \emph{n_thin}'th one. The number of saved
iterations thus equals \emph{(n_draw - n_burn) / n_thin}.}

\item{priors}{\code{bv_priors} object containing priors and their settings.
See \code{\link{bv_priors}}.}

\item{mh}{\code{bv_metropolis} object with settings regarding the acceptance
rate of the Metropolis-Hastings step. See \code{\link{bv_mh}}.}

\item{fcast}{\code{bv_fcast} object of forecast options set with
\code{\link{bv_fcast}}. May be set to \code{NULL} to skip forecasting.
Forecasts may also be calculated ex-post using \code{\link{predict.bvar}}.}

\item{irf}{\code{bv_irf} object with options regarding impulse responses and
forecast error variance decompositions. Set via \code{\link{bv_irf}} or
skipped when set to \code{NULL}. May also be computed ex-post using
\code{\link{irf.bvar}}.}

\item{verbose}{Logical scalar. Whether to print intermediate results and
progress.}

\item{...}{Not used.}
}
\value{
Returns a \code{bvar} object with the following elements:
\itemize{
  \item \code{beta} - Numeric array with saved draws from the posterior
  distribution of the VAR coefficients. See \code{\link{coef.bvar}}.
  \item \code{sigma} - Numeric array with saved draws from the posterior
  distribution of the model's VCOV-matrix. See \code{\link{vcov.bvar}}.
  \item \code{hyper} - Numeric matrix with saved draws from the posterior
  distributions of the hierarchical priors' hyperparameters.
  \item \code{ml} - Numeric vector with the values of the posterior
  marginalised likelihood (with respect to the hyperparameters).
  \item \code{optim} - List with outputs from \code{\link[stats]{optim}},
  which is used to find suitable starting values.
  \item \code{prior} - \code{bv_priors} object. See \code{\link{bv_priors}}.
  \item \code{call} - Call to the function. See \code{\link{match.call}}.
  \item \code{meta} - List with meta information such as number of variables,
  accepted draws, number of iterations, et cetera.
  \item \code{variables} - Character vector with the column names of
  \emph{data}. If missing variables are named consecutively - e.g.:
  \emph{var1}.
  \item \code{explanatories} - Character vector with the names of
  explanatory variables, i.e. the constant and lags. Formatting is akin to:
  \emph{FEDFUNDS-lag1}.
  \item \code{fcast} - \code{bvar_fcast} object with posterior forecast
  draws, quantiles as well as the forecast's setup from the \emph{fcast}
  argument.
  \item \code{irf} - \code{bvar_irf} object with posterior impulse response
  and their quantiles, forecast error variance decomposition draws, as well
  as the setup from the \emph{irf} argument.
}
}
\description{
Hierarchical Bayesian estimation of Vector Autoregression (VAR) models in
the fashion of Giannone, Primiceri and Lenza (2015).
Options for the prior settings and Metropolis-Hastings step are provided to
\emph{priors} and \emph{mh} arguments via the functions
\code{\link{bv_priors}} and \code{\link{bv_mh}}.
Several methods are available to analyse the resulting model, including
\code{\link{summary.bvar}}, \code{\link{plot.bvar}},
\code{\link{predict.bvar}} and \code{\link{irf.bvar}}.
}
\examples{
# Access a subset of the fred_qd dataset and transform it to be stationary
data("fred_qd")
data <- fred_qd[, c("CPIAUCSL", "UNRATE", "FEDFUNDS")]
data[5:nrow(data), 1] <- diff(log(data[, 1]), lag = 4) * 100
data <- data[5:nrow(data), ]

# Compute VAR using one lag and a ridiculously low number of draws
x <- bvar(data = data, lags = 1,
          n_draw = 500L, n_burn = 400L, n_thin = 2L, verbose = FALSE)

\donttest{
# Check out some of the outputs generated
plot(x)
predict(x)
plot(predict(x))
irf(x)
plot(irf(x))
}
}
\references{
Giannone, D. and Lenza, M. and Primiceri, G. E. (2015) Prior Selection for
  Vector Autoregressions. \emph{The Review of Economics and Statistics},
  \bold{97:2}, 436-451, \url{https://doi.org/10.1162/REST_a_00483}.
}
\seealso{
\code{\link{bv_priors}}; \code{\link{bv_mh}};
\code{\link{bv_fcast}}; \code{\link{bv_irf}};
\code{\link{predict.bvar}}; \code{\link{irf.bvar}}; \code{\link{plot.bvar}};
}
\author{
Nikolas Kuschnig, Lukas Vashold
}
\keyword{BVAR}
\keyword{VAR}
\keyword{hierarchical}
\keyword{macroeconomics}
\keyword{priors}
\keyword{vector-autoregression}
