% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fred_qd}
\alias{fred_qd}
\title{FRED-QD: Quarterly Database for Macroeconomic Research}
\format{A dataframe with 240 observations of 248 variables.}
\source{
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
fred_qd
}
\description{
FRED-QD is a large macroeconomic database designed for the empirical analysis
of "big data". They are updated in real-time through the FRED database.
The current version included is from April 2019 and contains quarterly
observations of many macroeconomic variables ranging from 1959Q1 to 2018Q4.
}
\references{
McCracken, M. W., & Ng, S. (2016): FRED-MD: A monthly database for macroeconomic research. Journal of Business & Economic Statistics, 34(4), 574-589.
}
\keyword{datasets}
