% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fred_transform}
\alias{fred_transform}
\title{FRED transformation helper}
\usage{
fred_transform(vars, type = c("full", "qd", "md"), code, lag = 1, scale = 100)
}
\arguments{
\item{vars}{Character vector.}

\item{type}{Character scalar. Specify whether to return both FRED-QD and
FRED-MD transformation codes (\code{"full"}) or subset to one (via
\code{"qd"} / \code{"md"}).}

\item{code}{Integer scalar.}

\item{lag}{Integer scalar. Number of lags to apply in the returned function
when taking differences. Defaults to 0.}

\item{scale}{Numeric scalar. Factor by which to scale up values after taking
differences in the returned function. Defaults to 100.}
}
\value{
Given \emph{vars} returns a named vector of transformation codes
for matching variables. Given \emph{code} returns a function to perform the
transformation. Note that this function does not perform input checks.
}
\description{
Retrieve and apply transformations from FRED-MD or FRED-QD. See
\code{\link{fred_qd}} for information on data and the details section for
information on available transformations. Note that the transformation
codes of FRED-MD and FRED-QD may differ.
}
\details{
The following transformation codes are available:
\enumerate{
  \item \code{1} - no transformation;
  \item \code{2} - first differences - \eqn{\delta x_t};
  \item \code{3} - second differences - \eqn{\delta^2 x_t};
  \item \code{4} - log transformation - \eqn{log(x_t)};
  \item \code{5} - log differences - \eqn{\delta^2 log(x_t)};
  \item \code{6} - percent change differences -
  \eqn{\delta x_t / x_{t-1} - 1};
}
}
\examples{
# Find transformation codes
fred_transform(c("RPI", "UNRATE"))

# Apply transformation #2 (first differences)
fred_transform(code = 2)(rnorm(10))
}
\seealso{
\code{\link{fred_md}}; \code{\link{fred_qd}};
}
