% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_plot.R, R/84_plot_hyper.R
\name{plot.bvar}
\alias{plot.bvar}
\alias{bv_plot}
\title{Hyperparameter plot}
\usage{
\method{plot}{bvar}(x, mar = c(2, 2, 2, 0.5), ...)

bv_plot(x, mar = c(2, 2, 2, 0.5), ...)
}
\arguments{
\item{x}{An \code{bvar}, obtained from \code{\link{bvar}}.}

\item{mar}{Numeric vector with margins for \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters for \code{\link[graphics]{par}}.}
}
\value{
Returns x invisibly.
}
\description{
A collection of diagnostic plots with traces and densities of the marginal
posterior likelihood and available hyperparameters. Minimum and maximum
values of the hyperparameters are plotted as darkgray dashed lines.
}
